# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Course API",
        "description": "API documentation for authentication, courses, interests, favorites, students, and teachers.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Application server"
        }
    ],
    "paths": {
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "operationId": "ad77fe138bcb384ad65d167d5c8d1cc0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Amine"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "amine@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "student",
                                        "enum": [
                                            "student",
                                            "teacher"
                                        ]
                                    },
                                    "interest_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2
                                        ]
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "name": "Amine",
                                "email": "amine@example.com",
                                "password": "password123",
                                "password_confirmation": "password123",
                                "role": "student",
                                "interest_ids": [
                                    1,
                                    2
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "User registered successfully.",
                                    "user": {
                                        "id": 1,
                                        "name": "Amine",
                                        "email": "amine@example.com",
                                        "role": "student"
                                    },
                                    "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9...",
                                    "token_type": "bearer"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "The given data was invalid.",
                                    "errors": {
                                        "email": [
                                            "The email has already been taken."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login and receive JWT token",
                "operationId": "bae68d1c55329ecb6e41f48285840c8f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "amine@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "email": "amine@example.com",
                                "password": "password123"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Login successful.",
                                    "user": {
                                        "id": 1,
                                        "name": "Amine",
                                        "email": "amine@example.com",
                                        "role": "student"
                                    },
                                    "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9...",
                                    "token_type": "bearer"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Invalid credentials."
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/forgot-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Generate reset password token",
                "operationId": "5113a3ca88126272e877fc19ea59499e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "amine@example.com"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "email": "amine@example.com"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reset token generated",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "If the email exists, reset token was generated.",
                                    "reset_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9...",
                                    "expires_in_minutes": 60
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "The given data was invalid.",
                                    "errors": {
                                        "email": [
                                            "The email field must be a valid email address."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/reset-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Reset password using reset token",
                "operationId": "eb313986b7dd1babb9c142612c89bab9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newpassword123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newpassword123"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9...",
                                "password": "newpassword123",
                                "password_confirmation": "newpassword123"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password reset successful",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Password reset successful.",
                                    "user": {
                                        "id": 1,
                                        "name": "Amine",
                                        "email": "amine@example.com",
                                        "role": "student"
                                    },
                                    "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9...",
                                    "token_type": "bearer"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid token or validation error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Invalid or expired reset token."
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}